requirements.txt
mir-eval==0.6
omegaconf==2.0.6
pytest==6.2.2
scikit-learn==0.23.2
tensorboard==2.4.0
tensorboardX==2.1
torch-summary==1.4.3
protobuf==3.20.1
matplotlib
umap-learn